+++ packages/ui-components/web/src/Components/FocusManager.tsx
@@ -1,3 +1,4 @@
+import type { KeyboardEventHandler } from 'react';
 import React, { Component, Fragment } from 'react';
 import FocusTrap from './FocusTrapReact';
 import type { Options as FocusTrapOptions } from 'focus-trap';
@@ -5,6 +6,7 @@ import type { Options as FocusTrapOptions } from 'focus-trap';
 interface FocusManagerProps {
   focusTrapOptions?: FocusTrapOptions;
   isMobile?: boolean;
+  onKeyDown?: KeyboardEventHandler;
 }
 
 class FocusManager extends Component<FocusManagerProps> {