+++ packages/ui-components/web/src/Components/Slider.tsx
@@ -3,8 +3,16 @@ import React, { useEffect, useState } from 'react';
 import classNames from 'classnames';
 import type { RichContentTheme } from 'wix-rich-content-common';
 import { mergeStyles } from 'wix-rich-content-common';
-import { ACTION_COLOR_CSS_VAR, BG_COLOR_CSS_VAR } from '../consts';
+import {
+  ACTION_COLOR_CSS_VAR,
+  BG_COLOR_CSS_VAR,
+  SLIDER_THUMB_VISIBILITY,
+  SLIDER_TRACK_SIZE,
+} from '../consts';
 import styles from '../../statics/styles/slider.scss';
+
+type ThumbVisibilityKeys = keyof typeof SLIDER_THUMB_VISIBILITY;
+type TrackSizeKeys = keyof typeof SLIDER_TRACK_SIZE;
 
 interface SliderProps {
   value: number;
@@ -12,15 +20,31 @@ interface SliderProps {
   max?: number;
   theme: RichContentTheme;
   onChange: (value: number) => void;
-  onSubmit: (value: number) => void;
+  onChangeCommitted?: (value: number) => void;
   dataHook?: string;
+  step?: string;
+  thumbVisibility?: typeof SLIDER_THUMB_VISIBILITY[ThumbVisibilityKeys];
+  trackSize?: typeof SLIDER_TRACK_SIZE[TrackSizeKeys];
+  height?: number;
   ariaProps?: InputHTMLAttributes<HTMLInputElement>;
   languageDir: string;
 }
 
 const Slider: FunctionComponent<SliderProps> = props => {
   const mergedStyles = mergeStyles({ styles, theme: props.theme });
-  const { min = 0, max = 10, onChange, dataHook, ariaProps, value, onSubmit, languageDir } = props;
+  const {
+    min = 0,
+    max = 10,
+    onChange,
+    dataHook,
+    ariaProps,
+    value,
+    onChangeCommitted,
+    languageDir,
+    thumbVisibility = SLIDER_THUMB_VISIBILITY.fixed,
+    trackSize = SLIDER_TRACK_SIZE.medium,
+    step,
+  } = props;
   const [fillPercentage, setFillPercentage] = useState(0);
   const track = {
     fill: ACTION_COLOR_CSS_VAR,
@@ -33,6 +57,8 @@ const Slider: FunctionComponent<SliderProps> = props => {
     } ${fillPercentage + 0.1}%)`,
   };
 
+  const isFixedThumb = thumbVisibility === SLIDER_THUMB_VISIBILITY.fixed;
+
   const onKeyUp = (event: KeyboardEvent) => {
     switch (event.key) {
       case 'ArrowUp':
@@ -43,7 +69,24 @@ const Slider: FunctionComponent<SliderProps> = props => {
       case 'End':
       case 'PageUp':
       case 'PageDown':
-        onSubmit((event.target as HTMLInputElement).valueAsNumber);
+        onChangeCommitted?.((event.target as HTMLInputElement).valueAsNumber);
+        break;
+      default:
+        return;
+    }
+  };
+
+  const onKeyDown = e => {
+    switch (e.key) {
+      case 'ArrowUp':
+      case 'ArrowDown':
+      case 'ArrowRight':
+      case 'ArrowLeft':
+      case 'Home':
+      case 'End':
+      case 'PageUp':
+      case 'PageDown':
+        onChange?.(e.target.valueAsNumber);
         break;
       default:
         return;
@@ -59,14 +102,23 @@ const Slider: FunctionComponent<SliderProps> = props => {
       {...ariaProps}
       tabIndex={0}
       type={'range'}
-      className={classNames(mergedStyles.slider, mergedStyles.wrapperSlider)}
+      className={classNames(
+        mergedStyles.slider,
+        mergedStyles.wrapperSlider,
+        mergedStyles[trackSize],
+        {
+          [mergedStyles.slider_fixed_thumb]: isFixedThumb,
+        }
+      )}
       data-hook={dataHook}
       onChange={e => onChange(e.target.valueAsNumber)}
       value={value}
       min={min}
       max={max}
-      onMouseUp={e => onSubmit((e.target as HTMLInputElement).valueAsNumber)}
+      step={step}
+      onMouseUp={e => onChangeCommitted?.((e.target as HTMLInputElement).valueAsNumber)}
       onKeyUp={onKeyUp}
+      onKeyDown={onKeyDown}
       style={bgStyle}
     />
   );